+++ Olympsis/Data/Models/OrganizationModels.swift
@@ -1,0 +1,8 @@
+//
+//  OrganizationModels.swift
+//  Olympsis
+//
+//  Created by Joel on 11/21/23.
+//
+
+import Foundation
