+++ Olympsis/Views/Viewers/Club/ClubTagView.swift
@@ -1,0 +1,18 @@
+//
+//  ClubTagView.swift
+//  Olympsis
+//
+//  Created by Joel on 11/19/23.
+//
+
+import SwiftUI
+
+struct ClubTagView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    ClubTagView()
+}
