--- OlympsisCompanion Watch App/OlympsisCompanionApp.swift
@@ -1,17 +1,0 @@
-//
-//  OlympsisCompanionApp.swift
-//  OlympsisCompanion Watch App
-//
-//  Created by Joel on 10/14/23.
-//
-
-import SwiftUI
-
-@main
-struct OlympsisCompanion_Watch_AppApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
