--- OlympsisCompanion Watch AppTests/OlympsisCompanion_Watch_AppTests.swift
@@ -1,36 +1,0 @@
-//
-//  OlympsisCompanion_Watch_AppTests.swift
-//  OlympsisCompanion Watch AppTests
-//
-//  Created by Joel on 10/14/23.
-//
-
-import XCTest
-@testable import OlympsisCompanion_Watch_App
-
-final class OlympsisCompanion_Watch_AppTests: XCTestCase {
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-        // Any test you write for XCTest can be annotated as throws and async.
-        // Mark your test throws to produce an unexpected failure when your test encounters an uncaught error.
-        // Tests marked async will run the test method on an arbitrary thread managed by the Swift runtime.
-    }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
